kosha-ai-sidecar/pyproject.toml
[project]
name = "kosha-ai-sidecar"
version = "0.1.0"
description = "Kosha KMS AI processing sidecar"
requires-python = ">=3.12"
dependencies = [
    "nats-py[nkeys]>=2.9.0",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-community>=0.3",
    "spacy>=3.8",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "minio>=7.2",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
